Data.ByteString.Lazy.compareLength returns Int64

+++ tests/Properties.hs
@@ -710,12 +710,12 @@ prop_all xs a = (all (== a) xs) == (L.all (== a) (pack xs))
 prop_maximum xs = (not (null xs)) ==> (maximum xs) == (L.maximum ( pack xs ))
 prop_minimum xs = (not (null xs)) ==> (minimum xs) == (L.minimum ( pack xs ))
 
-prop_compareLength1 xs  =  (L.pack xs         `L.compareLength` length xs) == EQ
-prop_compareLength2 xs c = (L.pack (xs ++ [c]) `L.compareLength` length xs) == GT
-prop_compareLength3 xs c = (L.pack xs `L.compareLength` length (xs ++ [c])) == LT
+prop_compareLength1 xs  =  (L.pack xs         `L.compareLength` fromIntegral (length xs)) == EQ
+prop_compareLength2 xs c = (L.pack (xs ++ [c]) `L.compareLength` fromIntegral (length xs)) == GT
+prop_compareLength3 xs c = (L.pack xs `L.compareLength` fromIntegral (length (xs ++ [c]))) == LT
 prop_compareLength4 xs c = ((L.pack xs `L.append` L.pack [c] `L.append` L.pack [undefined]) 
-                            `L.compareLength` length xs) == GT
-prop_compareLength5 xs l = L.compareLength xs l == compare (L.length xs) (fromIntegral l)
+                            `L.compareLength` fromIntegral (length xs)) == GT
+prop_compareLength5 xs l = L.compareLength xs l == compare (L.length xs) l
 
 prop_replicate1 c =
     forAll arbitrary $ \(Positive n) ->
